code clean up and bump numbering
