+++ bgpy/tests/engine_tests/engine_test_configs/internals/internal_config_013_aspa.py
@@ -4,7 +4,7 @@ from bgpy.as_graphs import ASGraphInfo, PeerLink
 from bgpy.as_graphs import CustomerProviderLink as CPLink
 from bgpy.shared.enums import ASNs, Prefixes
 from bgpy.simulation_engine import ASPA, BGP
-from bgpy.simulation_framework import AccidentalRouteLeak, ScenarioConfig    
+from bgpy.simulation_framework import AccidentalRouteLeak, ScenarioConfig
 from bgpy.tests.engine_tests.utils import EngineTestConfig
 
 r"""
@@ -23,22 +23,9 @@ ASPA Route Leak Attack Topology - Small Version
 - Shared path: AS 6 → AS 5 → ?? (ASPA validation decides who wins)
 """
 
-# Define AS Graph with proper attack topology
-as_graph_info = ASGraphInfo(
-    peer_links=frozenset({}),
-    customer_provider_links=frozenset({
-        CPLink(provider_asn=1, customer_asn=2),
-        CPLink(provider_asn=1, customer_asn=3),
-        CPLink(provider_asn=2, customer_asn=4),
-        CPLink(provider_asn=3, customer_asn=5),  # Shared path
-        CPLink(provider_asn=4, customer_asn=6),  # Victim path
-        CPLink(provider_asn=5, customer_asn=7),  # Attacker path
-    }),
-)
-
 # Define the ASPA Route Leak Scenario with corrected AS relationships
-internal_config_009_aspa = EngineTestConfig(
-    name="internal_config_009_aspa",
+internal_config_013_aspa = EngineTestConfig(
+    name="internal_config_013_aspa",
     desc=(
         "Tests ASPA validation in a small topology. "
         "Shows a case where ASPA protects paths, an attacker wins one path, "
@@ -46,18 +33,28 @@ internal_config_009_aspa = EngineTestConfig(
     ),
     scenario_config=ScenarioConfig(
         ScenarioCls=AccidentalRouteLeak,
-        BasePolicyCls=BGP,
+        BasePolicyCls=ASPA,
         override_victim_asns=frozenset({6}),
         override_attacker_asns=frozenset({7}),
-        hardcoded_asn_cls_dict=frozendict({
-            1: ASPA,  
-            2: ASPA,  
-            3: BGP,   
-            4: ASPA, 
-            5: BGP,   
-            6: ASPA,  # Victim
-            7: BGP,   # Attacker
-        }),
+        hardcoded_asn_cls_dict=frozendict(
+            {
+                3: BGP,
+                5: BGP,
+                7: BGP,  # Attacker
+            }
+        ),
     ),
-    as_graph_info=as_graph_info,
+    as_graph_info=ASGraphInfo(
+        peer_links=frozenset({}),
+        customer_provider_links=frozenset(
+            {
+                CPLink(provider_asn=1, customer_asn=2),
+                CPLink(provider_asn=1, customer_asn=3),
+                CPLink(provider_asn=2, customer_asn=4),
+                CPLink(provider_asn=3, customer_asn=5),  # Shared path
+                CPLink(provider_asn=4, customer_asn=6),  # Victim path
+                CPLink(provider_asn=5, customer_asn=7),  # Attacker path
+            }
+        ),
+    ),
 )
